Move base WPs to indico.web.views
