+++ indico/modules/categories/views.py
@@ -18,12 +18,12 @@ from __future__ import unicode_literals
 
 from markupsafe import escape
 
-from indico.legacy.webinterface.pages.base import WPDecorated, WPJinjaMixin
 from indico.legacy.webinterface.wcomponents import render_header
 from indico.modules.admin.views import WPAdmin
 from indico.util.i18n import _
 from indico.util.mathjax import MathjaxMixin
 from indico.web.breadcrumbs import render_breadcrumbs
+from indico.web.views import WPDecorated, WPJinjaMixin
 
 
 class WPManageUpcomingEvents(WPAdmin):
