+++ indico/modules/events/surveys/views.py
@@ -16,9 +16,9 @@
 
 from __future__ import unicode_literals
 
-from indico.legacy.webinterface.pages.base import WPJinjaMixin
 from indico.modules.events.management.views import WPEventManagement
 from indico.modules.events.views import WPConferenceDisplayBase, WPSimpleEventDisplayBase
+from indico.web.views import WPJinjaMixin
 
 
 class WPManageSurvey(WPEventManagement):
