+++ indico/modules/legal/views.py
@@ -16,10 +16,10 @@
 
 from __future__ import unicode_literals
 
-from indico.legacy.webinterface.pages.base import WPDecorated, WPJinjaMixin
 from indico.modules.admin.views import WPAdmin
 from indico.util.i18n import _
 from indico.web.breadcrumbs import render_breadcrumbs
+from indico.web.views import WPDecorated, WPJinjaMixin
 
 
 class WPLegalMixin:
