+++ indico/modules/news/views.py
@@ -16,9 +16,9 @@
 
 from __future__ import unicode_literals
 
-from indico.legacy.webinterface.pages.base import WPDecorated, WPJinjaMixin
 from indico.modules.admin.views import WPAdmin
 from indico.util.i18n import _
+from indico.web.views import WPDecorated, WPJinjaMixin
 
 
 class WPNews(WPJinjaMixin, WPDecorated):
